+++ docs/changelog.md
@@ -129,3 +129,4 @@ default scope due to a null FieldDescriptor.
 * Improvement: Simplified suggestion regex query.
 * Improvement: Added word boundary as prefix for words in suggestion handler instead of blank space.
 * Improvement: Instead of reading local solr schema as file to do schema check, now reading it as input stream.
+* Improvement: Removed invalid characters from file name and path on collection download